PICARD-714: Moved editable tag list implementation into own class
Added adding new item with double click.
