+++ picard/ui/options/interface_top_tags.py
@@ -17,16 +17,7 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-from PyQt5 import (
-    QtCore,
-    QtWidgets,
-)
-
 from picard import config
-from picard.util.tags import (
-    TAG_NAMES,
-    display_tag_name,
-)
 
 from picard.ui.options import (
     OptionsPage,
@@ -35,87 +26,6 @@ from picard.ui.options import (
 from picard.ui.ui_options_interface_top_tags import (
     Ui_InterfaceTopTagsOptionsPage,
 )
-
-
-class TagListModel(QtCore.QAbstractListModel):
-    def __init__(self, tags, parent=None):
-        super().__init__(parent)
-        self.tags = [(tag, display_tag_name(tag)) for tag in tags]
-
-    def rowCount(self, parent=QtCore.QModelIndex()):
-        return len(self.tags)
-
-    def data(self, index, role=QtCore.Qt.DisplayRole):
-        if not index.isValid() or role not in (QtCore.Qt.DisplayRole, QtCore.Qt.EditRole):
-            return None
-        i = index.row()
-        field = 1 if role == QtCore.Qt.DisplayRole else 0
-        try:
-            return self.tags[i][field]
-        except IndexError:
-            return None
-
-    def setData(self, index, value, role=QtCore.Qt.EditRole):
-        if not index.isValid() or role not in (QtCore.Qt.DisplayRole, QtCore.Qt.EditRole):
-            return False
-        i = index.row()
-        try:
-            if role == QtCore.Qt.EditRole:
-                display_name = display_tag_name(value) if value else value
-                self.tags[i] = (value, display_name)
-            elif role == QtCore.Qt.DisplayRole:
-                current = self.tags[i]
-                self.tags[i] = (current[0], value)
-            return True
-        except IndexError:
-            return False
-
-    def flags(self, index):
-        if index.isValid():
-            return (QtCore.Qt.ItemIsSelectable
-                | QtCore.Qt.ItemIsEditable
-                | QtCore.Qt.ItemIsDragEnabled
-                | QtCore.Qt.ItemIsEnabled
-                | QtCore.Qt.ItemNeverHasChildren)
-        else:
-            return QtCore.Qt.ItemIsDropEnabled
-
-    def insertRows(self, row, count, parent=QtCore.QModelIndex()):
-        super().beginInsertRows(parent, row, row + count - 1)
-        for i in range(count):
-            self.tags.insert(row, ("", ""))
-        super().endInsertRows()
-        return True
-
-    def removeRows(self, row, count, parent=QtCore.QModelIndex()):
-        super().beginRemoveRows(parent, row, row + count - 1)
-        self.tags = self.tags[:row] + self.tags[row + count:]
-        super().endRemoveRows()
-        return True
-
-    def supportedDragActions(self):
-        return QtCore.Qt.MoveAction
-
-    def supportedDropActions(self):
-        return QtCore.Qt.MoveAction
-
-    def clear(self):
-        self.beginResetModel()
-        self.tags = []
-        self.endResetModel()
-
-
-class TagItemDelegate(QtWidgets.QItemDelegate):
-    def createEditor(self, parent, option, index):
-        editor = QtWidgets.QLineEdit(parent)
-        completer = QtWidgets.QCompleter(TAG_NAMES.keys(), parent)
-
-        def complete(text):
-            parent.setFocus()
-
-        completer.activated.connect(complete)
-        editor.setCompleter(completer)
-        return editor
 
 
 class InterfaceTopTagsOptionsPage(OptionsPage):
@@ -144,20 +54,16 @@ class InterfaceTopTagsOptionsPage(OptionsPage):
 
     def load(self):
         tags = config.setting["metadatabox_top_tags"]
-        self._model = TagListModel(tags)
-        self.ui.top_tags_list.setItemDelegate(TagItemDelegate())
-        self.ui.top_tags_list.setModel(self._model)
+        self.ui.top_tags_list.update(tags)
 
     def save(self):
-        tags = []
-        for tag, name in self._model.tags:
-            tags.append(tag)
+        tags = list(self.ui.top_tags_list.tags)
         if tags != config.setting["metadatabox_top_tags"]:
             config.setting["metadatabox_top_tags"] = tags
             self.tagger.window.metadata_box.update()
 
     def restore_defaults(self):
-        self._model.clear()
+        self.ui.top_tags_list.clear()
         super().restore_defaults()
