+++ picard/ui/taglistview.py
@@ -1,0 +1,151 @@
+# -*- coding: utf-8 -*-
+#
+# Picard, the next-generation MusicBrainz tagger
+# Copyright (C) 2019 Philipp Wolfer
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+
+from PyQt5 import (
+    QtCore,
+    QtWidgets,
+)
+
+from picard.util.tags import (
+    TAG_NAMES,
+    display_tag_name,
+)
+
+
+class EditableTagListView(QtWidgets.QListView):
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        model = TagListModel()
+        self.setModel(model)
+        self.setItemDelegate(TagItemDelegate())
+
+    def mouseDoubleClickEvent(self, event):
+        index = self.indexAt(QtCore.QPoint(event.x(), event.y()))
+        if index.isValid():
+            super().mouseDoubleClickEvent(event)
+        else:
+            self.add_empty_row()
+
+    def add_tag(self, tag=""):
+        model = self.model()
+        row = model.rowCount()
+        model.insertRow(row)
+        index = model.createIndex(row, 0)
+        model.setData(index, tag)
+        return index
+
+    def clear(self):
+        self.model().update([])
+
+    def update(self, tags):
+        self.model().update(tags)
+
+    @property
+    def tags(self):
+        return self.model().tags
+
+    def add_empty_row(self):
+        index = self.add_tag()
+        self.setCurrentIndex(index)
+        self.edit(index)
+
+
+class TagListModel(QtCore.QAbstractListModel):
+    def __init__(self, tags=[], parent=None):
+        super().__init__(parent)
+        self._tags = [(tag, display_tag_name(tag)) for tag in tags]
+
+    def rowCount(self, parent=QtCore.QModelIndex()):
+        return len(self._tags)
+
+    def data(self, index, role=QtCore.Qt.DisplayRole):
+        if not index.isValid() or role not in (QtCore.Qt.DisplayRole, QtCore.Qt.EditRole):
+            return None
+        i = index.row()
+        field = 1 if role == QtCore.Qt.DisplayRole else 0
+        try:
+            return self._tags[i][field]
+        except IndexError:
+            return None
+
+    def setData(self, index, value, role=QtCore.Qt.EditRole):
+        if not index.isValid() or role not in (QtCore.Qt.DisplayRole, QtCore.Qt.EditRole):
+            return False
+        i = index.row()
+        try:
+            if role == QtCore.Qt.EditRole:
+                display_name = display_tag_name(value) if value else value
+                self._tags[i] = (value, display_name)
+            elif role == QtCore.Qt.DisplayRole:
+                current = self._tags[i]
+                self._tags[i] = (current[0], value)
+            return True
+        except IndexError:
+            return False
+
+    def flags(self, index):
+        if index.isValid():
+            return (QtCore.Qt.ItemIsSelectable
+                | QtCore.Qt.ItemIsEditable
+                | QtCore.Qt.ItemIsDragEnabled
+                | QtCore.Qt.ItemIsEnabled
+                | QtCore.Qt.ItemNeverHasChildren)
+        else:
+            return QtCore.Qt.ItemIsDropEnabled
+
+    def insertRows(self, row, count, parent=QtCore.QModelIndex()):
+        super().beginInsertRows(parent, row, row + count - 1)
+        for i in range(count):
+            self._tags.insert(row, ("", ""))
+        super().endInsertRows()
+        return True
+
+    def removeRows(self, row, count, parent=QtCore.QModelIndex()):
+        super().beginRemoveRows(parent, row, row + count - 1)
+        self._tags = self._tags[:row] + self._tags[row + count:]
+        super().endRemoveRows()
+        return True
+
+    def supportedDragActions(self):
+        return QtCore.Qt.MoveAction
+
+    def supportedDropActions(self):
+        return QtCore.Qt.MoveAction
+
+    def update(self, tags):
+        self.beginResetModel()
+        self._tags = [(tag, display_tag_name(tag)) for tag in tags]
+        self.endResetModel()
+
+    @property
+    def tags(self):
+        return (t[0] for t in self._tags)
+
+
+class TagItemDelegate(QtWidgets.QItemDelegate):
+    def createEditor(self, parent, option, index):
+        editor = QtWidgets.QLineEdit(parent)
+        completer = QtWidgets.QCompleter(TAG_NAMES.keys(), parent)
+
+        def complete(text):
+            parent.setFocus()
+
+        completer.activated.connect(complete)
+        editor.setCompleter(completer)
+        return editor
