+++ picard/ui/ui_options_interface_top_tags.py
@@ -18,7 +18,7 @@ class Ui_InterfaceTopTagsOptionsPage(object):
         self.label = QtWidgets.QLabel(InterfaceTopTagsOptionsPage)
         self.label.setObjectName("label")
         self.vboxlayout.addWidget(self.label)
-        self.top_tags_list = QtWidgets.QListView(InterfaceTopTagsOptionsPage)
+        self.top_tags_list = EditableTagListView(InterfaceTopTagsOptionsPage)
         self.top_tags_list.setDragEnabled(True)
         self.top_tags_list.setDragDropMode(QtWidgets.QAbstractItemView.InternalMove)
         self.top_tags_list.setObjectName("top_tags_list")
@@ -30,3 +30,4 @@ class Ui_InterfaceTopTagsOptionsPage(object):
     def retranslateUi(self, InterfaceTopTagsOptionsPage):
         _translate = QtCore.QCoreApplication.translate
         self.label.setText(_("Show the below tags above all other tags in the metadata view"))
+from picard.ui.taglistview import EditableTagListView
