+++ ui/options_interface_top_tags.ui
@@ -34,7 +34,7 @@
     </widget>
    </item>
    <item>
-    <widget class="QListView" name="top_tags_list">
+    <widget class="EditableTagListView" name="top_tags_list">
      <property name="dragEnabled">
       <bool>true</bool>
      </property>
@@ -45,6 +45,13 @@
    </item>
   </layout>
  </widget>
+ <customwidgets>
+  <customwidget>
+   <class>EditableTagListView</class>
+   <extends>QListView</extends>
+   <header>picard.ui.taglistview</header>
+  </customwidget>
+ </customwidgets>
  <resources/>
  <connections/>
 </ui>
